Not-perfect but rather simple and doing-what-we-want fix for #931: everytime the user accesses the WMI `shares` page its shares will be checked. This will trigger the short URL requesting. As the user is expected to view the WMI page to get the short URL for transmitting, this ensures the URL request can be triggered manually by the user.
\\

This will obviously trigger more checks than really needed, but parallel requests are already avoided by `check()` if URL is already set.

\\

BTW, I added a new internal event emitted when a new URL is set. This will trigger a notification on the WMI side, asking the user to reload the page. This is not as good as users/groups pages do (row-based ajax update), but the missing part will be implemented after the `implement#new_machine_page` branch is merged into `develop`. No need to duplicate the work now for such a small feature, as the branch is coming soon.

\\

BTW2: I refreshed / cleaned the `simplesharing` WMI event handlers, and notably removed the extension enabled test, which seems double-work, as events are not emited if the extension is not enabled…